webapp/requirements.txt
fastapi
uvicorn[standard]
redis
python-multipart
pydantic
torch
torchvision
Pillow
opencv-python-headless
